requirements.txt
django==1.11.1
djangorestframework==3.6.3
pygments==2.2.0
django-rest-swagger==2.1.2
django-cors-headers==2.1.0
django-extensions==1.7.9
